refactor: 상수 패키지 수정
