+++ tomcat/src/main/java/org/apache/coyote/http/response/StatusLine.java
@@ -2,8 +2,8 @@ package org.apache.coyote.http.response;
 
 import org.apache.coyote.http.HttpVersion;
 
-import static org.apache.coyote.http.Constants.CRLF;
-import static org.apache.coyote.http.Constants.SPACE;
+import static org.apache.coyote.util.Constants.CRLF;
+import static org.apache.coyote.util.Constants.SPACE;
 
 public class StatusLine {
