+++ tomcat/src/main/java/org/apache/coyote/util/StringUtils.java
@@ -1,7 +1,5 @@
 package org.apache.coyote.util;
 
-import org.apache.coyote.http.Constants;
-
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
